Minor PR fixes
